fix(admin-ui): rename cache-refresh to jans-link #1185
