+++ admin-ui/openapi-merge.json
@@ -13,7 +13,7 @@
         "inputURL": "https://raw.githubusercontent.com/JanssenProject/jans/main/jans-config-api/plugins/docs/user-mgt-plugin-swagger.yaml"
       },
       {
-        "inputURL": "https://raw.githubusercontent.com/JanssenProject/jans/main/jans-config-api/plugins/docs/cache-refresh-plugin-swagger.yaml"
+        "inputURL": "https://raw.githubusercontent.com/JanssenProject/jans/main/jans-config-api/plugins/docs/jans-link-plugin-swagger.yaml"
       }
     ], 
     "output": "./configApiSpecs.yaml"
