+++ admin-ui/plugins.config.json
@@ -28,7 +28,7 @@
     "metadataFile": "./fido/plugin-metadata"
   },
   {
-    "key": "cache-refresh",
-    "metadataFile": "./cache-refresh/plugin-metadata"
+    "key": "jans-link",
+    "metadataFile": "./jans-link/plugin-metadata"
   }
 ]
