+++ admin-ui/plugins/jans-link/components/Tabs/InumDBServerTab.js
@@ -12,7 +12,7 @@ import BindPasswordModal from '../CacheRefresh/BindPasswordModal'
 import * as Yup from 'yup'
 import GluuCommitFooter from 'Routes/Apps/Gluu/GluuCommitFooter'
 import { isEmpty } from 'lodash'
-import { putCacheRefreshConfiguration } from 'Plugins/cache-refresh/redux/features/CacheRefreshSlice'
+import { putCacheRefreshConfiguration } from 'Plugins/jans-link/redux/features/CacheRefreshSlice'
 import GluuCommitDialog from 'Routes/Apps/Gluu/GluuCommitDialog'
 import { buildPayload } from 'Utils/PermChecker'
 
@@ -96,7 +96,7 @@ const InumDBServerTab = () => {
     toggleAudit()
 
     buildPayload(userAction, userMessage, {
-      cacheRefreshConfiguration: {
+      jansLinkConfiguration: {
         ...cacheRefreshConfiguration,
         targetConfig: {
           ...formik.values.targetConfig,
